+++ lib/neoid/config.rb
@@ -1,0 +1,6 @@
+module Neoid
+  class Config
+    attr_accessor :enable_subrefs
+    attr_accessor :enable_per_model_indexes
+  end
+end