+++ lib/neoid/middleware.rb
@@ -5,10 +5,12 @@ module Ndoid
     end
 
     def call(env)
-      old, Thread.current[:neoid_enabled] = Thread.current[:neoid_enabled], true
+      old_enabled, Thread.current[:neoid_enabled] = Thread.current[:neoid_enabled], true
+      old_batch, Thread.current[:neoid_current_batch] = Thread.current[:neoid_current_batch], nil
       @app.call(env)
     ensure
-      Thread.current[:neoid_enabled] = old
+      Thread.current[:neoid_enabled] = old_enabled
+      Thread.current[:neoid_current_batch] = old_batch
     end
   end
 end