+++ lib/neoid/node.rb
@@ -1,15 +1,46 @@
 module Neoid
   module Node
+    def self.from_hash(hash)
+      node = Neography::Node.new(hash)
+      node.neo_server = Neoid.db
+      node
+    end
+
     module ClassMethods
+      attr_accessor :neo_subref_node
+
       def neo_subref_rel_type
         @_neo_subref_rel_type ||= "#{self.name.tableize}_subref"
       end
+
       def neo_subref_node_rel_type
         @_neo_subref_node_rel_type ||= self.name.tableize
       end
-      
+
+      def delete_command
+        :delete_node
+      end
+
+      def neo_model_index
+        return nil unless Neoid.config.enable_per_model_indexes
+
+        Neoid::logger.info "Node#neo_model_index #{neo_subref_rel_type}"
+
+        gremlin_query = <<-GREMLIN
+          g.createManualIndex(neo_model_index_name, Vertex.class);
+        GREMLIN
+
+        # Neoid.logger.info "subref query:\n#{gremlin_query}"
+
+        script_vars = { neo_model_index_name: neo_model_index_name }
+
+        Neoid.execute_script_or_add_to_batch gremlin_query, script_vars
+      end
+
       def neo_subref_node
-        @_neo_subref_node ||= begin
+        return nil unless Neoid.config.enable_subrefs
+
+        @neo_subref_node ||= begin
           Neoid::logger.info "Node#neo_subref_node #{neo_subref_rel_type}"
 
           gremlin_query = <<-GREMLIN
@@ -22,17 +53,24 @@ module Neoid
               g.addEdge(g.v(0), subref, neo_subref_rel_type);
             }
 
-            g.createManualIndex(neo_index_name, Vertex.class);
-
             subref
           GREMLIN
 
-          Neoid.logger.info "subref query:\n#{gremlin_query}"
+          # Neoid.logger.info "subref query:\n#{gremlin_query}"
 
-          node = Neography::Node.load(Neoid.db.execute_script(gremlin_query, neo_subref_rel_type: neo_subref_rel_type, name: self.name, neo_index_name: self.neo_index_name))
+          script_vars = {
+            neo_subref_rel_type: neo_subref_rel_type,
+            name: self.name
+          }
 
-          node
+          Neoid.execute_script_or_add_to_batch gremlin_query, script_vars do |value|
+            Neoid::Node.from_hash(value)
+          end
         end
+      end
+
+      def reset_neo_subref_node
+        @neo_subref_node = nil
       end
 
       def neo_search(term, options = {})
@@ -42,48 +80,72 @@ module Neoid
     
     module InstanceMethods
       def neo_find_by_id
-        Neoid::logger.info "Node#neo_find_by_id #{self.class.neo_index_name} #{self.id}"
-        Neoid.db.get_node_index(self.class.neo_index_name, :ar_id, self.id)
-      rescue Neography::NotFoundException
-        nil
+        # Neoid::logger.info "Node#neo_find_by_id #{self.class.neo_index_name} #{self.id}"
+        node = Neoid.db.get_node_auto_index(Neoid::UNIQUE_ID_KEY, self.neo_unique_id)
+        node.present? ? Neoid::Node.from_hash(node[0]) : nil
       end
       
-      def neo_create
+      def _neo_save
         return unless Neoid.enabled?
-        
-        data = self.to_neo.merge(ar_type: self.class.name, ar_id: self.id)
+
+        data = self.to_neo.merge(ar_type: self.class.name, ar_id: self.id, Neoid::UNIQUE_ID_KEY => self.neo_unique_id)
         data.reject! { |k, v| v.nil? }
-        
-        node = Neography::Node.create(data)
-        
-        retires = 2
-        begin
-          Neography::Relationship.create(
-            self.class.neo_subref_node_rel_type,
-            self.class.neo_subref_node,
-            node
+
+        gremlin_query = <<-GREMLIN
+          idx = g.idx('node_auto_index');
+          q = null;
+          if (idx) q = idx.get(unique_id_key, unique_id);
+
+          node = null;
+          if (q && q.hasNext()) {
+            node = q.next();
+            node_data.each {
+              if (node.getProperty(it.key) != it.value) {
+                node.setProperty(it.key, it.value);
+              }
+            }
+          } else {
+            node = g.addVertex(node_data);
+            if (enable_subrefs) g.addEdge(g.v(subref_id), node, neo_subref_rel_type);
+
+            if (enable_model_index) g.idx(neo_model_index_name).put('ar_id', node.ar_id, node);
+          }
+
+          node
+        GREMLIN
+
+         script_vars = {
+          unique_id_key: Neoid::UNIQUE_ID_KEY,
+          node_data: data,
+          unique_id: self.neo_unique_id,
+          enable_subrefs: Neoid.config.enable_subrefs,
+          enable_model_index: Neoid.config.enable_per_model_indexes && self.class.neoid_config.enable_model_index
+        }
+
+        if Neoid.config.enable_subrefs
+          script_vars.update(
+            subref_id: self.class.neo_subref_node.neo_id,
+            neo_subref_rel_type: self.class.neo_subref_rel_type
           )
-        rescue
-          # something must've happened to the cached subref node, reset and retry
-          @_neo_subref_node = nil
-          retires -= 1
-          retry if retires > 0
         end
 
-        Neoid.db.add_node_to_index(self.class.neo_index_name, :ar_id, self.id, node)
+        if Neoid.config.enable_per_model_indexes && self.class.neoid_config.enable_model_index
+          script_vars.update(
+            neo_model_index_name: self.class.neo_model_index_name
+          )
+        end
 
-        Neoid::logger.info "Node#neo_create #{self.class.name} #{self.id}, index = #{self.class.neo_index_name}"
-        
-        neo_search_index
+        Neoid::logger.info "Node#neo_save #{self.class.name} #{self.id}"
+
+        node = Neoid.execute_script_or_add_to_batch(gremlin_query, script_vars) do |value|
+          @_neo_representation = Neoid::Node.from_hash(value)
+        end.then do |result|
+          neo_search_index
+        end
 
         node
       end
-      
-      def neo_update
-        Neoid.db.set_node_properties(neo_node, self.to_neo)
-        neo_search_index
-      end
-      
+
       def neo_search_index
         return if self.class.neoid_config.search_options.blank? || (
           self.class.neoid_config.search_options.index_fields.blank? &&
@@ -113,21 +175,12 @@ module Neoid
         end
       end
       
-      def neo_load(node)
-        Neography::Node.load(node)
+      def neo_load(hash)
+        Neoid::Node.from_hash(hash)
       end
       
       def neo_node
         _neo_representation
-      end
-    
-      def neo_destroy
-        return unless neo_node
-        Neoid.db.remove_node_from_index(DEFAULT_FULLTEXT_SEARCH_INDEX_NAME, neo_node)
-
-        Neoid.db.remove_node_from_index(self.class.neo_index_name, neo_node)
-        neo_node.del
-        _reset_neo_representation
       end
 
       def neo_after_relationship_remove(relationship)
@@ -135,7 +188,7 @@ module Neoid
       end
 
       def neo_before_relationship_through_remove(record)
-        rel_model, foreign_key_of_owner, foreign_key_of_record = Neoid.config[:relationship_meta_data][self.class.name.to_s][record.class.name.to_s]
+        rel_model, foreign_key_of_owner, foreign_key_of_record = Neoid::Relationship.meta_data[self.class.name.to_s][record.class.name.to_s]
         rel_model = rel_model.to_s.constantize
         @__neo_temp_rels ||= {}
         @__neo_temp_rels[record] = rel_model.where(foreign_key_of_owner => self.id, foreign_key_of_record => record.id).first
@@ -152,10 +205,6 @@ module Neoid
       receiver.extend         ClassMethods
       receiver.send :include, InstanceMethods
       Neoid.node_models << receiver
-      
-      receiver.after_create :neo_create
-      receiver.after_update :neo_update
-      receiver.after_destroy :neo_destroy
     end
   end
 end